mysite/requirement.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.4
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
python-http-client==3.3.7
requests==2.28.1
sendgrid==6.9.7
soupsieve==2.3.2.post1
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
sendgrid==6.9.7
urllib3==1.26.13
yarl==1.8.2
django-simple-captcha==0.5.17